Extracted commands into new file.
