+++ src/VSCodeExtension/src/telemetry.ts
@@ -29,6 +29,7 @@ export namespace EventNames {
     export const lspReady = "lsp-ready";
     export const lspStopped = "lsp-stopped";
     export const error = "error";
+    export const commandExecuted = "command-executed";
 }
 
 // @ts-ignore
